[common,thread] Store thread function on the stack
The thread function was stored as static variable inside Thread constructor.
But it requires guards for thread-safe initialization of static instances. These
guards are part of C++ stdlib in gcc. Disabling them with
-fno-threadsafe-statics compiler options may lead to undefined behavior here or
other places.

The only place where we can safely store the function is beginning of the thread
stack.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>